Organized types of voting page and favicon pic.
