+++ together-wherever/src/hooks/useDestinationVoting.ts
@@ -2,7 +2,8 @@
 
 import { useState, useEffect } from "react";
 import { useRouter, useParams } from "next/navigation";
-import { DestinationInterface, TripDetailsInterface } from "@/utils/voting-page";
+
+import { DestinationInterface, TripDetailsInterface } from "@/utils/types";
 import { fetchVotingPageData } from "@/fetcher/voting-page";
 
 export function useDestinationVoting() {
